LayoutTests/fast/repaint/no-animation-outside-viewport-expected.txt

Test that animated gif outside viewport does not trigger repaint.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS repaintRects is ""
PASS successfullyParsed is true

TEST COMPLETE
